# perf: limit initSchema calls from openapi.IsNamespaceScoped 