Avoid calling initSchema from openapi.IsNamespaceScoped when possible. Work done in https://github.com/kubernetes-sigs/kustomize/pull/4152 introduced a precomputed namespace scope map based on the default built-in schema. This commit extends that work by avoiding calls to initSchema when a resource is not found in the precomputed map and the default built-in schema is in use. In those cases, there is no benefit to calling initSchema since the precomputed map is exactly what will be calculated by parsing the default built-in schema.

This PR resolves some of the concern in #4569, but the PR does not fix the issue completely. Specifically, this PR does not address this [earlier comment](https://github.com/kubernetes-sigs/kustomize/issues/4569#issuecomment-1244692695) from @KnVerey:

> We should add a --proto option to kustomize openapi fetch and also allow proto-formatted schemas in the openapi field.

The PR is split into two commits, one for a new benchmark and one for the performance improvement. Using the benchmark for analysis...

Before:

```
$ go test ./kyaml/openapi -bench BenchmarkPrecomputed -benchmem -run '^$'
goos: windows
goarch: amd64
pkg: sigs.k8s.io/kustomize/kyaml/openapi
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
BenchmarkPrecomputedIsNamespaceScoped/unknown_resource-8                      22          52757164 ns/op        46676575 B/op     339637 allocs/op
BenchmarkPrecomputedIsNamespaceScoped/namespace_scoped-8                22427852                52.83 ns/op            0 B/op          0 allocs/op
BenchmarkPrecomputedIsNamespaceScoped/cluster_scoped-8                  21984786                54.13 ns/op            0 B/op          0 allocs/op
PASS
ok      sigs.k8s.io/kustomize/kyaml/openapi     5.260s
```

After:

```
$ go test ./kyaml/openapi -bench BenchmarkPrecomputed -benchmem -run '^$'
goos: windows
goarch: amd64
pkg: sigs.k8s.io/kustomize/kyaml/openapi
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
BenchmarkPrecomputedIsNamespaceScoped/cluster_scoped-8          18558757                53.92 ns/op            0 B/op          0 allocs/op
BenchmarkPrecomputedIsNamespaceScoped/unknown_resource-8        15135080                76.36 ns/op            0 B/op          0 allocs/op
BenchmarkPrecomputedIsNamespaceScoped/namespace_scoped-8        22635700                52.51 ns/op            0 B/op          0 allocs/op
PASS
ok      sigs.k8s.io/kustomize/kyaml/openapi     4.102s
```