@ephesused 2023-03-14
It turns out there's more complexity in the processing, and I want to take a little time to think about this one before coding further.

Thanks to the test coverage - particularly [api/krusty's TestCustomOpenApiFieldFromComponentWithOverlays](https://github.com/kubernetes-sigs/kustomize/blob/master/api/krusty/openapicustomschema_test.go#L441) - it's clear this optimization introduces a bug. What's not clear is how to retain the current behavior and provide the optimization.

At first, it would seem simple - just delay the initSchema loads until the point where the schema truly is needed. The problem is that by the time the schema truly is needed, the earlier state has been lost. In effect, the question centers around what the initSchema calls would have done for earlier calls to NewGvk when the initSchema invocations were skipped. That's certainly possible, but it's more complex than I had thought it would be.

Explanation...

For the code at master, the TestCustomOpenApiFieldFromComponentWithOverlays flow does something like so:

(api/krusty/kustomizer.go) Kustomizer.Run calls openapi.SetSchema
(api/internal/target/kusttarget.go) KustTarget.accumulateDirectory calls openapi.SetSchema

Later, during accumulation:

(api/resmap.factory.go) Factory.FromFile chains down to...
(kyaml/resid/gvk.go) NewGvk calls...
(kyaml/openapi/openapi.go) IsCertainlyClusterScoped calls...
(kyaml/openapi/openapi.go) initSchema - at this point that's against the default

Later, still during accumulation:

(api/internal/target/kusttarget.go) KustTarget.accumulateDirectory calls openapi.SetSchema based on the defined openapi path

And a little later:

(kyaml/resid/gvk.go) NewGvk leads to...
(kyaml/openapi/openapi.go) initSchema - at this point that's against the defined openapi path

The result is that the stored data now has both the default and the defined openapi path.

The change I proposed eliminates the initSchema call from certain NewGvk calls. As a result, the default schema might not load. That absence leads to the testcase failure. Currently, NewGvk has a side effect of loading the default schema when the Gvk is not known. It turns out that the default schema load is a requirement for certain cases - and TestCustomOpenApiFieldFromComponentWithOverlays is one of them.

I believe the reason TestCustomOpenApiFieldFromComponentWithOverlays is affected is that [com.github.openshift.api.apps.v1.DeploymentConfigSpec is defined via a reference to io.k8s.api.core.v1.PodTemplateSpec](https://github.com/kubernetes-sigs/kustomize/blob/master/api/krusty/testdata/openshiftschema.json#L30-L37). In the absence of the default schema load, the reference is not present. In turn, that affects the results of the patch.