@ephesused 2023-09-01
> Looks like a test failure with `TestCustomOpenApiFieldFromComponentWithOverlays`.

Thanks - I'll try to take a look at that later today.