@ephesused 2023-09-01
That was a bit embarrassing - I don't know how I managed to mess that up four months ago.

@natasha41575, I believe the last commit should resolve the issue. I'm never able to run all the tests cleanly in my environment, but the ones I checked look good.

For example:

```
$ git log --pretty=oneline | head -1
7e104198440938beca50a1fa17b8ccad5479fd96 fix: correct openapi.go's schemaNotParsed value
$ go test ./api/krusty -run TestCustomOpenApi -count 1
ok      sigs.k8s.io/kustomize/api/krusty        0.709s
$ go test ./kyaml/openapi/... -count 1
?       sigs.k8s.io/kustomize/kyaml/openapi/kubernetesapi       [no test files]
?       sigs.k8s.io/kustomize/kyaml/openapi/kubernetesapi/v1_21_2       [no test files]
?       sigs.k8s.io/kustomize/kyaml/openapi/kustomizationapi    [no test files]
ok      sigs.k8s.io/kustomize/kyaml/openapi     0.241s
$
```